# Conflicts:
#	enrollment-service/src/main/java/com/pbl/elearning/enrollment/repository/EnrollmentRepository.java
#	web/src/main/java/com/pbl/elearning/web/endpoint/enrollment/EnrollmentController.java

+++ enrollment-service/src/main/java/com/pbl/elearning/enrollment/payload/response/EnrollmentReportResponse.java
@@ -19,7 +19,7 @@ public class EnrollmentReportResponse {
     private String courseTitle;
     private String enrollmentStatus;
     private Double progressPercentage;
-    private Integer totalWatchTimeMinutes;
+    private Double totalWatchTimeMinutes;
     private LocalDateTime enrollmentDate;
     private LocalDateTime completionDate;
     private LocalDateTime lastAccessedAt;
